+++ spec/www/spec/db-tx-value-bindings-test.js
@@ -71,7 +71,7 @@ var mytests = function() {
         if (isWebSql) {
           return window.openDatabase(name, "1.0", "Demo", DEFAULT_SIZE);
         } else {
-          return window.sqlitePlugin.openDatabase({name: name});
+          return window.sqlitePlugin.openDatabase(name, '1.0', 'Demo', -1);
         }
       }